src/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
motor>=3.3.2
beanie>=1.23.6
pydantic>=2.5.2 